// src/screens/tools/shapes/BorderRadius/types.ts
export type TBorderRadiusConfig = {
  topLeft: number;
  topRight: number;
  bottomRight: number;
  bottomLeft: number;
  unit: "px" | "%" | "rem";
  elementWidth: number;
  elementHeight: number;
  backgroundColor: string;
  useAdvanced: boolean;
  topLeftH: number;
  topLeftV: number;
  topRightH: number;
  topRightV: number;
  bottomRightH: number;
  bottomRightV: number;
  bottomLeftH: number;
  bottomLeftV: number;
};